requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
sympy>=1.12
tokenizers>=0.13.0
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
accelerate>=0.20.0
einops>=0.6.0